# services/api/openapi.yaml
openapi: 3.1.0
info:
  title: Graphview API
  version: 0.4.0
  description: Runnable Graphview API contract for graph, source, ingestion, proposal, review, and import/export flows.
paths:
...
  /sources/{source_id}:
    patch:
      summary: Update a source.
      parameters:
        - name: source_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated source.
    delete:
      summary: Delete a source.
      parameters:
        - name: source_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted source.
  /ingestion-runs:
    get:
      summary: List ingestion runs and current stages.
      responses:
        "200":
          description: Ingestion run list response.
    post:
      summary: Ingest text, markdown, URL, or PDF content and create reviewable proposals.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestionCreate"
      responses:
        "201":
          description: Created source, ingestion run, proposals, and embeddings.
        "422":
          description: Missing required source content.
        "502":
          description: Source fetch or extraction failed.
...
components:
  schemas:
...
    IngestionCreate:
      type: object
      required: [kind, title]
      properties:
        kind:
          type: string
          enum: [text, markdown, url, pdf]
        title:
          type: string
        content:
          type: string
        uri:
          type: string
        content_base64:
          type: string
        proposal_limit:
          type: integer
          minimum: 1
          maximum: 12
          default: 4